+++ changelog.md
@@ -1,6 +1,10 @@
 - Added a NixOS flake
 - The Linux AppImage can optionally add itself to autostart and the application menu at onboarding
-- Improved YouTube title parsing
+- Added YouTube and YouTube Music support on Linux. The browser must report the page URL to MPRIS
+  for this to work
+- Improved YouTube title parsing for Linux and Android
+- Added an option to discard and "Fetch missing album" for a regex edit. This is useful for radio
+  apps that put the station name in the album field, for example.
 - Option to hide the loved status in Discord Rich Presence
 - Merged the scrobble notifications with persistent notification on Android
 - Bug fixes